Backport - Images: Thumbnails are deleted after change of image title on Azure storage

+++ Telerik.Sitefinity.Azure/BlobStorage/AzureBlobStorageProvider.cs
@@ -58,6 +58,8 @@ namespace Telerik.Sitefinity.Azure.BlobStorage
 
             blob.Properties.ContentType = content.MimeType;
 
+            blob.Metadata[nameof(IBlobContent.FileId)] = content.FileId.ToString();
+
             return blob.OpenWrite();
         }
 
@@ -87,13 +89,22 @@ namespace Telerik.Sitefinity.Azure.BlobStorage
         public override void Delete(IBlobContentLocation content)
         {
             CloudBlob blob = this.GetBlob(content);
+
             try
             {
-                var requestOptions = new Storage.Blob.BlobRequestOptions()
+                if (blob.Exists())
                 {
-                    RetryPolicy = new Storage.RetryPolicies.NoRetry() // RetryPolicies.Retry(1, TimeSpan.FromSeconds(1))
-                };
-                blob.DeleteIfExists(DeleteSnapshotsOption.None, null, requestOptions, null);
+                    blob.FetchAttributes();
+                    bool hasFileId = blob.Metadata.ContainsKey(nameof(IBlobContentLocation.FileId));
+                    if (!hasFileId || (hasFileId && blob.Metadata[nameof(IBlobContentLocation.FileId)].Equals(content.FileId.ToString())))
+                    {
+                        var requestOptions = new Storage.Blob.BlobRequestOptions()
+                        {
+                            RetryPolicy = new Storage.RetryPolicies.NoRetry() // RetryPolicies.Retry(1, TimeSpan.FromSeconds(1))
+                        };
+                        blob.DeleteIfExists(DeleteSnapshotsOption.None, null, requestOptions, null);
+                    }
+                }
             }
             catch (Exception e)
             {
@@ -160,6 +171,9 @@ namespace Telerik.Sitefinity.Azure.BlobStorage
 
                 this.SetProperties(destination, this.GetProperties(source));
             }
+
+            destBlob.Metadata[nameof(IBlobContentLocation.FileId)] = source.FileId.ToString();
+            destBlob.SetMetadata();
         }
 
         /// <inheritdoc />
